# src/migration-coverage/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Migration Coverage API
  version: 1.0.0
  description: |
    Comprehensive API for tracking and validating database migration coverage across clinical, business, communications, and technical data domains.

    This API provides real-time visibility into migration progress, data integrity validation, and comprehensive reporting for large-scale database migrations.

  contact:
    name: Migration Coverage API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.migration-coverage.example.com
    description: Production server
paths:
...
  /validation/results/{id}:
    get:
      tags:
        - Validation
      summary: Get validation results
      description: Returns validation results for a specific job
      parameters:
        - name: id
          in: path
          required: true
          description: Validation job ID
          schema:
            type: string
            format: uuid
        - name: includeDetails
          in: query
          description: Include detailed validation results
          schema:
            type: boolean
            default: false
        - name: format
          in: query
          description: Response format
          schema:
            type: string
            enum: [json, report]
            default: json
      responses:
        '200':
          description: Validation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResults'
            text/markdown:
              schema:
                type: string
                description: Validation report in Markdown format
        '404':
          description: Validation job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ValidationResults:
      type: object
      required:
        - jobId
        - status
        - startTime
        - summary
        - integrityChecks
        - recommendations
        - timestamp
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
          nullable: true
        duration:
          type: number
          nullable: true
        summary:
          $ref: '#/components/schemas/ValidationSummary'
        integrityChecks:
          type: object
          required:
            - total
            - passed
            - failed
          properties:
            total:
              type: integer
              minimum: 0
            passed:
              type: integer
              minimum: 0
            failed:
              type: integer
              minimum: 0
        recommendations:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
    ValidationSummary:
      type: object
      required:
        - totalRules
        - passedRules
        - failedRules
        - criticalFailures
        - warningCount
        - overallScore
        - executionTime
      properties:
        totalRules:
          type: integer
          minimum: 0
        passedRules:
          type: integer
          minimum: 0
        failedRules:
          type: integer
          minimum: 0
        criticalFailures:
          type: integer
          minimum: 0
        warningCount:
          type: integer
          minimum: 0
        overallScore:
          type: number
          minimum: 0
          maximum: 100
        executionTime:
          type: number
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          description: Request identifier for debugging
        path:
          type: string
          description: Request path
        method:
          type: string
          description: HTTP method
...